Move `PythonBinary` from `python.binaries` to `core.util_rules.system_binaries` (#14609)
Builds off of https://github.com/pantsbuild/pants/pull/14601.

In a followup, I'll propose moving the remainder of `python.binaries` to a new location.

[ci skip-rust]
[ci skip-build-wheels]
